+++ src/CoreEx/Hosting/TimerHostedServiceBase.cs
@@ -25,8 +25,12 @@ namespace CoreEx.Hosting
     {
         private static readonly Random _random = new();
 
-        private readonly TimerHostedServiceHealthCheck? _healthCheck;
+#if NET9_0_OR_GREATER
+        private readonly System.Threading.Lock _lock = new();
+#else
         private readonly object _lock = new();
+#endif
+        private readonly TimerHostedServiceHealthCheck? _healthCheck;
         private TimerHostedServiceStatus _status = TimerHostedServiceStatus.Initialized;
         private string? _name;
         private CancellationTokenSource? _cts;